Backend/requirements.txt
langchain==0.3.21
langchain-cli==0.0.36
langchain-community==0.3.20
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-pinecone==0.2.3
langchain-tests==0.3.16
langchain-text-splitters==0.3.7
langdetect==1.0.9
langgraph==0.3.21
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.60
langserve==0.3.1
langsmith==0.3.18
numpy==1.26.4
Flask==3.1.0
openai==1.68.2
pinecone==5.4.2
pinecone-client==6.0.0
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
flask-cors
gunicorn
pycryptodome


